Добавил UserService и UserServiceImpl. Реализовал во втором часть методов. Поправил Валидатор.

+++ src/main/java/ru/kata/spring/boot_security/demo/repositories/UsersRepository.java
@@ -6,6 +6,8 @@ import ru.kata.spring.boot_security.demo.models.User;
 
 import java.util.Optional;
 @Repository
-public interface UsersRepository extends JpaRepository<User, Integer> {
+public interface UsersRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
+
+
 }
